refactor: models
