+++ app/entrypoints/autonomous.py
@@ -8,7 +8,8 @@ from sqlmodel import Session, select
 
 from app.config.config import config
 from app.core.ai import execute_agent
-from app.models.db import Agent, AgentQuota, get_engine, init_db
+from app.models.agent import Agent, AgentQuota
+from app.models.db import get_engine, init_db
 
 logger = logging.getLogger(__name__)
